Ensure float32

+++ train_TEDD1104.py
@@ -138,7 +138,7 @@ def inference_tedd1104(
                 load_in_4bit=True,
                 bnb_4bit_use_double_quant=True,
                 bnb_4bit_quant_type="nf4",
-                bnb_4bit_compute_dtype=torch.bfloat16
+                bnb_4bit_compute_dtype=torch.float16
                 if torch.cuda.is_bf16_supported() and not training_args.fp16_full_eval
                 else torch.float32,
             )
